requirements.txt
asgiref>=3.2.3
django>=3.0.3
pytz>=2019.3
sqlparse>=0.3.0
djangorestframework>=3.16.1
markdown>=3.8.2
django-filter>=25.1
pytest>=8.4.2
factory_boy>=3.3.3
pytest-factoryboy>=2.8.1
pytest-django>=4.11.1
requests>=2.30.0
responses>=0.25.8
